requirements.txt
selenium==3.141.0
pytesseract==0.3.4
xlrd==1.2.0
Pillow==8.2.0
requests==2.22.0